# Snapshot in-memory chunks on shutdown for faster restarts