--- a/tsdb/docs/format/README.md
+++ b/tsdb/docs/format/README.md
@@ -5,3 +5,4 @@
 * [Head Chunks](head_chunks.md)
 * [Tombstones](tombstones.md)
 * [Wal](wal.md)
+* [Memory Snapshot](memory_snapshot.md)