--- a/tsdb/encoding/encoding.go
+++ b/tsdb/encoding/encoding.go
@@ -17,6 +17,7 @@ import (
 	"encoding/binary"
 	"hash"
 	"hash/crc32"
+	"math"
 	"unsafe"
 
 	"github.com/dennwc/varint"
@@ -40,6 +41,7 @@ func (e *Encbuf) Len() int    { return len(e.B) }
 
 func (e *Encbuf) PutString(s string) { e.B = append(e.B, s...) }
 func (e *Encbuf) PutByte(c byte)     { e.B = append(e.B, c) }
+func (e *Encbuf) PutBytes(b []byte)  { e.B = append(e.B, b...) }
 
 func (e *Encbuf) PutBE32int(x int)      { e.PutBE32(uint32(x)) }
 func (e *Encbuf) PutUvarint32(x uint32) { e.PutUvarint64(uint64(x)) }
@@ -54,6 +56,10 @@ func (e *Encbuf) PutBE32(x uint32) {
 func (e *Encbuf) PutBE64(x uint64) {
 	binary.BigEndian.PutUint64(e.C[:], x)
 	e.B = append(e.B, e.C[:8]...)
+}
+
+func (e *Encbuf) PutBEFloat64(x float64) {
+	e.PutBE64(math.Float64bits(x))
 }
 
 func (e *Encbuf) PutUvarint64(x uint64) {
@@ -71,6 +77,12 @@ func (e *Encbuf) PutUvarintStr(s string) {
 	b := *(*[]byte)(unsafe.Pointer(&s))
 	e.PutUvarint(len(b))
 	e.PutString(s)
+}
+
+// PutUvarintBytes writes a a variable length byte buffer.
+func (e *Encbuf) PutUvarintBytes(b []byte) {
+	e.PutUvarint(len(b))
+	e.PutBytes(b)
 }
 
 // PutHash appends a hash over the buffers current contents to the buffer.
@@ -247,6 +259,10 @@ func (d *Decbuf) Be64() uint64 {
 	x := binary.BigEndian.Uint64(d.B)
 	d.B = d.B[8:]
 	return x
+}
+
+func (d *Decbuf) Be64Float64() float64 {
+	return math.Float64frombits(d.Be64())
 }
 
 func (d *Decbuf) Be32() uint32 {