This PR is based on this design doc: https://docs.google.com/document/d/1c90BmhErybHH94hoGNsgJC9pyaUTU3dgHoyD3dm6Kw4/edit?usp=sharing (the doc is a little outdated after some discussions here)

Snapshot happens during shutdown and not at regular intervals to avoid write amplification.

The snapshot name is of the format `chunk_snapshot.<last WAL segment>.<offset in the segment>`. The offset in the WAL segment gives us more precision on where to continue the snapshot and avoids unnecessary replay of entire WAL segment.

**Format of the snapshot**
Starts with series records. Each series gets 1 record.

```
┌────────────────────────────┬────────────────────────────┐
│      Record Type <byte>    │    Series Ref <uint64>     │
├────────────────────────────┴────────────────────────────┤
│                Number of  Labels <uvarint>              │
├────────────────────────────┬────────────────────────────┤
│    len(name_1) <uvarint>   │      name_1 <bytes>        │
├────────────────────────────┼────────────────────────────┤
│    len(val_1) <uvarint>    │      val_1 <bytes>         │
├────────────────────────────┴────────────────────────────┤
│                          . . .                          │
├────────────────────────────┬────────────────────────────┤
│    len(name_N) <uvarint>   │      name_N <bytes>        │
├────────────────────────────┼────────────────────────────┤
│    len(val_N) <uvarint>    │      val_N <bytes>         │
├────────────────────────────┴────────────────────────────┤
│                   Chunk Range <int64>                   │
├─────────────────────────────────────────────────────────┤
│                  Chunk Exists <uvarint>                 │
│  # 1 if head chunk exists, 0 otherwise to detect a nil  |
|  # chunk. Below fields exists only when it's 1 here.    |
├────────────────────────────┬────────────────────────────┤
│      Chunk Mint <int64>    │     Chunk Maxt <int64>     │
├────────────────────────────┴────────────────────────────┤
│                  Chunk Encoding <byte>                  │
├────────────────────────────┬────────────────────────────┤
│     len(Chunk) <uvarint>   │     Chunk <bytes>          │
├────────────────────────────┼────────────────────────────┤
|   sampleBuf[0].t <int64>   |  sampleBuf[0].v <float64>  | 
├────────────────────────────┼────────────────────────────┤
|   sampleBuf[1].t <int64>   |  sampleBuf[1].v <float64>  | 
├────────────────────────────┼────────────────────────────┤
|   sampleBuf[2].t <int64>   |  sampleBuf[2].v <float64>  | 
├────────────────────────────┼────────────────────────────┤
|   sampleBuf[3].t <int64>   |  sampleBuf[3].v <float64>  | 
└────────────────────────────┴────────────────────────────┘
```

Ends with a single record for tombstones. The encoded tombstones uses the same encoding as tombstone file in blocks.

```
┌─────────────────────────────────────────────────────────────────┐
│                        Record Type <byte>                       │
├───────────────────────────────────┬─────────────────────────────┤
│ len(Encoded Tombstones) <uvarint> │ Encoded Tombstones <bytes>  │
└───────────────────────────────────┴─────────────────────────────┘
```