@codesome 2020-05-20
There were some bugs in the replay part in terms of handing the file, will benchmark again after fixing them. But from what I could gather with this:

Snapshotting chunks ~1.5M series takes ~3.6s, lets take 4s for any anomalies. So that is like ~30s for 10M series.

After running for 2hrs, replay of master took 2m34s. Excluding the mmap replay of 14s, that naively translates to ~11.6s for 10m worth of WAL data.

Replaying the chunk snapshot took around 11s with snapshot interval being 10m. As there was some bug in code, taking 10m of WAL replay from above, replaying entire data with snapshots would be 11+14+11.6=36.6s roughly in the worst case. That is 4-5x faster replay, and expect it to be >5x in general case.

I forgot to check the size of snapshots.

Before I dive into fixing the issues, how do these numbers look for a trade-off? @brian-brazil