@codesome 2020-05-20
Additionally, replay time won't increase linearly (or linear with very small factor) with partial chunks like WAL replay would. This was tested only for 2h of WAL data, but in extended use, I guess it goes upto 3h or more. So in that case, the difference would be >5x.