@codesome 2020-05-20
I compared only the replay times here whereas the turnaround time is the one that matters. With <4s for taking a snapshot (with most hovering around 3s), I see not much change with the speedup even for turnaround time :)