@brian-brazil 2020-05-20
To summaries:
90s for 30M series, or an additional ~15% duty cycle on I/O for 10m snapshots. That's a fair bit.

The benchmark was for ~250k/s. For 1M/s, we'd be looking at 2h of WAL replay taking ~616s with the current mmap code.

Replaying just the chunk snapshot for 30M series would be ~220s plus ~46s for the WAL replay at 1M/s so ~266s in all. So the saving here is only about 2x for very large Prometheus servers, and we're still over a minute.

I'm not sure this complexity is worth it given all the additional disk writes it'd cause. I'd prefer to keep things simple to allow for some future solution that'd get us under a minute for large servers.