@codesome 2020-05-21
> get us under a minute for large servers.

Might be possible for only replays, but for turn around time that sounds very ambitious unless we can write out the data to disk faster or m-map partial chunks (unlikely) :) I will think of more ways