@codesome 2020-05-26
From the way I see it, for the sake of some arbitrary target we should not block some optimization which is possible right now. The tradeoff seems to be good.

Additionally, the code and the logic are well separated from the rest of the Head block logic and I don't see this being a complex addition. It is also safe to downgrade without any issues as it will fallback to the old replay which makes it easy to upgrade to any new techniques which hits the target in the future. This could also live behind a flag for now with disabled by default.