@brian-brazil 2020-05-26
I can see that argument, but the duty cycle and additional disk writes still concern me. Do you have a rough idea of what the overall impact of the writes is on top of the WAL/mmap/compactions we already do? Write amplification is something we should keep an eye on, and I think this could double it.