@brancz 2020-05-26
I agree we need to measure and decide based on what's measured, but I don't think anyone suggested anything contrary to that.