@codesome 2020-05-27
Here are some calculations (for a single series)

Considering a time window of 2h, and 15s scrape interval, the calculations are based on how many chunks will be written to disk for the chunks created in these 2h in it's entire lifecycle.

**Current writes**

Chunks written in 2h to disk = 4 (from mmapping) + 4 (to write to blocks) = 8

I am not totally sure about the compression ratio between WAL and XORChunk, but considering that a chunk in WAL would be twice the size than XORChunk. So here it is 8 + 8 = 16. I will add +2 for the checkpoints (or should I not?). So now it's 12+2 = 14.

If we consider that a block goes through at least 2 compactions in it's lifecycle, then writing those 4 chunks twice = 16 + (4+4) = 24

**Writes from partial chunks**

Considering 15m snapshot interval, we can say it writes 0.5 chunks at 15m then 1 chunks at 30m. As the cycle continues, it would be 4*(0.5+1) = 6 chunks for 2 hrs.

**Additional writes**

6 additional chunks compared to 24 chunks currently, that is 25% increase in writes.

I have not considered block index writes and >2 compactions which will happen, so increase in writes would be less than <20% I believe.

Did I miss something?