@brian-brazil 2020-05-27
> considering that a chunk in WAL would be twice the size than XORChunk

It's going to be about 3x with compression, presuming a conservative 2 bytes/samples in a chunk.

> Considering 15m snapshot interval

Why the switch from the 10m used previously?

My math is in terms of a single sample which is conservatively ~2 bytes. In terms of full blocks we'll currently write it out 9 times through compaction, plus another 6 bytes for a compressed WAL which is the same as 3 writes. So our current write amplification is around 12.

Mmaping I believe works out as no more than one additional write so that'd bring it to 13.

Snapshots of incomplete chunks every 10m could write a sample out 12 times if the chunk is never completed, which is plausible with a 60s interval. Halve that for the average case as later samples in the 2h get written fewer times, so 6. Overall impact is 13->19 so roughly adding ~46% more write amplification.

Is shaving off half of restart time worth reducing SSD lifetime by a third?


Doing the math for a Samsung 860 EVO 500GB with 200k/s (i.e. 15d retention just fits) which quotes 300TB of writes that takes the official supported lifetime from ~22 months to ~15 months. Now I'd presume it'll last a bit longer than that, but it's not as if it's 22 years versus 15 years where it'd be clear that it wouldn't be a problem.