@codesome 2020-06-06
Just a thought: given the write amplification with regular snapshotting, one of the ways to use this snapshotting can be to snapshot only when Prometheus is exiting. This way, a graceful restart will be fast, but not crashes. 

With graceful restarts being very few in number, I expect not much impact on write amplification.