@brian-brazil 2020-06-08
That would reduce the disk writes, but then we'd be slowing down shutdown and the code would be exercised less so more likely a bug will slip in. I'm a bit wary of that, and keep in mind that the shutdown time would now needed to be added in to the restart time.