@codesome 2020-06-09
> but then we'd be slowing down shutdown

> and keep in mind that the shutdown time would now needed to be added in to the restart time.

But the overall turnaround time is still faster. Maybe even faster than previous calculation because with snapshotting at the end, the WAL to be replayed is 0, and snapshotting takes less time than replaying that extra WAL in the old calculation.

Some recalculation with snapshot at the end with the data gathered in this comment https://github.com/prometheus/prometheus/pull/7229#issuecomment-631465164

For the turn around time we only have writing the snapshot + the replay of mmap and snapshot (no WAL)

For 1.5M series for 2h of data

* Master = 2m34s for replay (hence turnaround time)
* With Snapshotting = 4s (writing snapshot) + 14s (mmap replay) + 11s (snapshot replay) + 0s (WAL replay) = 29s

So that is ~80% reduction; better than before because the earlier calculation was based on the worst case of having to replay some WAL.

As we will have more than 2h of WAL, the reduction is going to be more.

> the code would be exercised less so more likely a bug will slip in

The bug can be fixed :) It is similar to WAL replay - it is also exercised only when Prometheus starts so more likely a bug will slip, but we still have it.